backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.2
sympy==1.13.3
torch==2.2.2+cpu
torchvision==0.17.2+cpu
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
